+++ pypsa/optimization/constraints.py
@@ -630,16 +630,19 @@ def define_operational_constraints_for_committables(
     # linearized approximation because committable can partly start up and shut down
     start_up_cost = c.da.start_up_cost.sel(name=com_i)
     shut_down_cost = c.da.shut_down_cost.sel(name=com_i)
-    cost_equal = (start_up_cost == shut_down_cost).values
+    cost_equal_da = start_up_cost == shut_down_cost
+    if "snapshot" in cost_equal_da.dims:
+        cost_equal_da = cost_equal_da.all("snapshot")
+    cost_equal = cost_equal_da.values
 
     # only valid additional constraints if start up costs equal to shut down costs
     if n._linearized_uc and not cost_equal.all():
         logger.warning(
             "The linear relaxation of the unit commitment cannot be "
             "tightened for all generators since the start up costs "
-            "are not equal to the shut down costs. Proceed with the "
-            "linear relaxation without the tightening by additional "
-            "constraints for these. This might result in a longer "
+            "are not equal to the shut down costs in all snapshots. "
+            "Proceed with the linear relaxation without the tightening by "
+            "additional constraints for these. This might result in a longer "
             "solving time."
         )